// backend/src/common/enums/role.enum.ts
export enum UserRole {
  SUPER_ADMIN = 'SUPER_ADMIN',
  ADMIN_RESTAURANT = 'ADMIN_RESTAURANT',
  MANAGER = 'MANAGER',
  SERVEUR = 'SERVEUR',
  CUISINE = 'CUISINE',
  BAR = 'BAR',
  LIVREUR = 'LIVREUR', // Chauffeur/Livreur pour les livraisons
  CAISSIER = 'CAISSIER', // Pour l'encaissement
  STOCK = 'STOCK', // Gestionnaire de stock
  CLIENT = 'CLIENT',
}